navbar setup. polished css. restructured components directory
