+++ app/client/src/App.js
@@ -1,27 +1,16 @@
 import React, { Component } from 'react';
 import './App.css';
 import { connect } from 'react-redux';
-import UploadDropzone from './components/uploadDropzone.js';
-import UploadResult from './components/uploadResult.js';
+import UploadDropzone from './components/home/uploadDropzone.js';
 
 class App extends Component {
-
-
-    constructor(props) {
-        super(props);
-        this.state = {
-
-        };
-    }
-
-
 
     render() {
         return (
             <div className="App container">
                 <h1>Welcome to RIS!</h1>
                 <div><UploadDropzone /></div>
-                <div>{this.props.names ? <UploadResult names={this.props.names}/> : null}</div>
+                {/*}<div>{this.props.names ? <UploadResult names={this.props.names}/> : null}</div>*/}
             </div>
         );
     }
