+++ app/client/src/components/home/uploadDropzone.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import './uploadDropzone.css';
 import DragAndDrop from './dragAndDrop.js'
 import { connect } from 'react-redux';
-import { uploadImage } from '../redux/analysis/action.js';
+import { uploadImage } from '../../redux/analysis/action.js';
 import { FaFileUpload } from 'react-icons/fa';
 
 class UploadDropzone extends Component {
@@ -52,8 +52,13 @@ class UploadDropzone extends Component {
                         </DragAndDrop>
                     </div>
                     <div>
-                        <input type="file" id="imageUpload" onChange={this.handleChangeImageUpload} accept="image/*"/>
-                        <button type="submit" className="btn btn-dark" onClick={(e) => {this.handleClickImageUpload(e)}}>Upload</button>
+                    <div className="input-group mb-3 browse-button-grp">
+                        <div className="custom-file">
+                            <input type="file" className="custom-file-input" id="imageUpload" onChange={this.handleChangeImageUpload} accept="image/*"/>
+                            <label className="custom-file-label" htmlFor="imageUpload">{this.state.newImage ? this.state.newImage.name : "Choose file"}</label>
+                        </div>
+                    </div>
+                        <button type="submit" className="btn btn-dark browse-button-grp" onClick={(e) => {this.handleClickImageUpload(e)}}>Upload</button>
                     </div>
                 </form>
             </div>
