+++ app/client/src/components/navbar.js
@@ -1,0 +1,30 @@
+import React, { Component } from 'react';
+import { Link } from 'react-router-dom';
+import './navbar.css';
+
+class Navbar extends Component {
+
+    render() {
+        return (
+            <nav className="navbar navbar-expand-lg navbar-dark bg-dark">
+                <Link to="/" className="navbar-brand">RIS</Link>
+                <button className="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+                    <span className="navbar-toggler-icon"></span>
+                </button>
+
+                <div className="collapse navbar-collapse" id="navbarSupportedContent">
+                    <ul className="navbar-nav mr-auto">
+                        <li className="nav-item active">
+                            <Link to="/home" className="nav-link">Home</Link>
+                        </li>
+                        <li className="nav-item">
+                            <Link to="/about" className="nav-link">About</Link>
+                        </li>
+                    </ul>
+                </div>
+            </nav>
+        );
+    }
+}
+
+export default Navbar;
